+++ lib/nav-bar-styles/style-3-bottom-nav-bar.widget.dart
@@ -30,7 +30,8 @@ class BottomNavStyle3 extends StatelessWidget {
       this.horizontalPadding,
       this.navBarCurve});
 
-  Widget _buildItem(PersistentBottomNavBarItem item, bool isSelected, double height) {
+  Widget _buildItem(
+      PersistentBottomNavBarItem item, bool isSelected, double height) {
     return AnimatedContainer(
       width: 100.0,
       height: height / 1.0,
@@ -49,8 +50,12 @@ class BottomNavStyle3 extends StatelessWidget {
                 data: IconThemeData(
                     size: iconSize,
                     color: isSelected
-                        ? (item.activeContentColor == null ? item.activeColor : item.activeContentColor)
-                        : item.inactiveColor == null ? item.activeColor : item.inactiveColor),
+                        ? (item.activeContentColor == null
+                            ? item.activeColor
+                            : item.activeContentColor)
+                        : item.inactiveColor == null
+                            ? item.activeColor
+                            : item.inactiveColor),
                 child: item.icon,
               ),
             ),
@@ -60,7 +65,11 @@ class BottomNavStyle3 extends StatelessWidget {
                 type: MaterialType.transparency,
                 child: DefaultTextStyle.merge(
                   style: TextStyle(
-                      color: isSelected ? (item.activeContentColor == null ? item.activeColor : item.activeContentColor) : item.inactiveColor,
+                      color: isSelected
+                          ? (item.activeContentColor == null
+                              ? item.activeColor
+                              : item.activeContentColor)
+                          : item.inactiveColor,
                       fontWeight: FontWeight.w400,
                       fontSize: item.titleFontSize),
                   child: FittedBox(child: Text(item.title)),
@@ -76,7 +85,8 @@ class BottomNavStyle3 extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     Color selectedItemActiveColor = items[selectedIndex].activeColor;
-    double itemWidth = (MediaQuery.of(context).size.width / items.length) - ((MediaQuery.of(context).size.width * 0.03) / 3);
+    double itemWidth = (MediaQuery.of(context).size.width / items.length) -
+        ((MediaQuery.of(context).size.width * 0.03) / 3);
     return Container(
       decoration: getNavBarDecoration(
         navBarCurve: this.navBarCurve,
@@ -87,14 +97,21 @@ class BottomNavStyle3 extends StatelessWidget {
         child: BackdropFilter(
           filter: ImageFilter.blur(sigmaX: 3.0, sigmaY: 3.0),
           child: Container(
-            color: getBackgroundColor(context, items, backgroundColor, selectedIndex),
+            color: getBackgroundColor(
+                context, items, backgroundColor, selectedIndex),
             child: Container(
               width: double.infinity,
               height: this.navBarHeight,
               padding: EdgeInsets.only(
-                  left: this.horizontalPadding == null ? MediaQuery.of(context).size.width * 0.05 : this.horizontalPadding,
-                  right: this.horizontalPadding == null ? MediaQuery.of(context).size.width * 0.05 : this.horizontalPadding,
-                  bottom: this.bottomPadding == null ? this.navBarHeight * 0.1 : this.bottomPadding),
+                  left: this.horizontalPadding == null
+                      ? MediaQuery.of(context).size.width * 0.05
+                      : this.horizontalPadding,
+                  right: this.horizontalPadding == null
+                      ? MediaQuery.of(context).size.width * 0.05
+                      : this.horizontalPadding,
+                  bottom: this.bottomPadding == null
+                      ? this.navBarHeight * 0.1
+                      : this.bottomPadding),
               child: Column(
                 children: <Widget>[
                   Row(
@@ -104,7 +121,8 @@ class BottomNavStyle3 extends StatelessWidget {
                         color: Colors.transparent,
                         width: selectedIndex == 0
                             ? MediaQuery.of(context).size.width * 0.0
-                            : itemWidth * (selectedIndex) - MediaQuery.of(context).size.width * 0.025,
+                            : itemWidth * (selectedIndex) -
+                                MediaQuery.of(context).size.width * 0.025,
                         height: 4.0,
                       ),
                       Flexible(
@@ -140,7 +158,8 @@ class BottomNavStyle3 extends StatelessWidget {
                               },
                               child: Container(
                                 color: Colors.transparent,
-                                child: _buildItem(item, selectedIndex == index, this.navBarHeight),
+                                child: _buildItem(item, selectedIndex == index,
+                                    this.navBarHeight),
                               ),
                             ),
                           );