+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.3.1] - 2020-04-29
 
-- Implemented Android back button handling.
+- Implemented handling of the Android back button.
 - Fixed the issue where the app would not close at all on Android back button press.
 - Updated navigation bar height to give it the default platform look.
 - Updated styles to fix the issue where a tap would not be registered.